flo/deb: optimize ART
--Turn off ART watchdog

There is a time limit for optimization of apk.
Some apps like play services takes a lot of time,
so it doesn't gets optimized properly due to time limit
and tries to optimize on each both.It also happens when it's updated.
By disabling watchdog we let the apk to get optimized completely.

Change-Id: I6a856b25f71b103be8307a112ff04d40d82a1a93